requirements.txt
numpy
pandas
scipy
scipy-stubs
matplotlib
attrs